Move IDs methods in an ID allocator module
