+++ src/python/ensembl/io/genomio/gff3/simplifier.py
@@ -38,6 +38,7 @@ from importlib_resources import files
 import ensembl.io.genomio.data.gff3
 from ensembl.io.genomio.utils.json_utils import get_json
 from .extract_annotation import FunctionalAnnotations
+from .id_allocator import IDAllocator
 
 
 class Records(list):
@@ -73,10 +74,7 @@ class GFFSimplifier:
     allow_pseudogene_with_CDS = False
     exclude_seq_regions: List = []
     fail_types: Dict[str, int] = {}
-    validate_gene_id = True
-    min_id_length = 8
-    stable_id_prefix = None
-    current_stable_id_number: int = 0
+    ids = IDAllocator()
 
     def __init__(self, genome_path: Optional[PathLike] = None, make_missing_stable_ids: bool = False):
         biotypes_json = files(ensembl.io.genomio.data.gff3) / "biotypes.json"
@@ -87,7 +85,7 @@ class GFFSimplifier:
         if genome_path:
             with Path(genome_path).open("r") as genome_fh:
                 self.genome = json.load(genome_fh)
-        self.make_missing_stable_ids: bool = make_missing_stable_ids
+        self.ids.make_missing_stable_ids = make_missing_stable_ids
 
     def simpler_gff3(self, in_gff_path: PathLike) -> None:
         """Loads a GFF3 from INSDC and rewrites it in a simpler version, whilst also writing a
@@ -246,7 +244,7 @@ class GFFSimplifier:
             return feat
 
         # Generate ID if needed and add it to the functional annotation
-        feat.id = self.normalize_gene_id(feat)
+        feat.id = self.ids.normalize_gene_id(feat)
         self.annotations.add_feature(feat, "transposable_element")
         feat.qualifiers = {"ID": feat.id}
 
@@ -287,7 +285,7 @@ class GFFSimplifier:
         """
 
         # New gene ID
-        gene.id = self.normalize_gene_id(gene)
+        gene.id = self.ids.normalize_gene_id(gene)
 
         # Gene with no subfeatures: need to create a transcript at least
         if len(gene.sub_features) == 0:
@@ -330,7 +328,7 @@ class GFFSimplifier:
 
         # PSEUDOGENE CDS IDs
         if gene.type == "pseudogene" and self.allow_pseudogene_with_CDS:
-            self.normalize_pseudogene_cds_id(gene)
+            self.ids.normalize_pseudogene_cds_id(gene)
 
         return gene
 
@@ -357,7 +355,7 @@ class GFFSimplifier:
 
             # New transcript ID
             transcript_number = count + 1
-            transcript.id = self.normalize_transcript_id(gene.id, transcript_number)
+            transcript.id = self.ids.normalize_transcript_id(gene.id, transcript_number)
 
             transcript = self.format_gene_segments(transcript)
 
@@ -387,7 +385,7 @@ class GFFSimplifier:
                     feat.qualifiers["source"] = old_exon_qualifiers["source"]
             elif feat.type == "CDS":
                 # New CDS ID
-                feat.id = self.normalize_cds_id(feat.id)
+                feat.id = self.ids.normalize_cds_id(feat.id)
                 if feat.id in ("", gene.id, transcript.id):
                     feat.id = f"{transcript.id}_cds"
             else:
@@ -453,7 +451,7 @@ class GFFSimplifier:
         gene = SeqFeature(cds.location, type=gene_type)
         gene.qualifiers["source"] = cds.qualifiers["source"]
         gene.sub_features = [transcript]
-        gene.id = self.generate_stable_id()
+        gene.id = self.ids.generate_stable_id()
 
         return gene
 
@@ -661,144 +659,3 @@ class GFFSimplifier:
             transcript.sub_features = new_subfeats
             gene_subfeats.append(transcript)
         gene.sub_features = gene_subfeats
-
-# IDS
-    def normalize_gene_id(self, gene: SeqFeature) -> str:
-        """Remove any unnecessary prefixes around the gene ID.
-
-        Generate a new stable id if it is not recognized as valid.
-
-        Args:
-            gene: Gene feature to normalize.
-
-        Returns:
-            A normalized gene id.
-
-        """
-
-        prefixes = ["gene-", "gene:"]
-        new_gene_id = self.remove_prefixes(gene.id, prefixes)
-
-        # In case the gene id is not valid, use the GeneID
-        if not self.valid_id(new_gene_id):
-            logging.warning(f"Gene id is not valid: {new_gene_id}")
-            qual = gene.qualifiers
-            if "Dbxref" in qual:
-                for xref in qual["Dbxref"]:
-                    (db, value) = xref.split(":")
-                    if db == "GeneID":
-                        new_gene_id = f"{db}_{value}"
-                        logging.debug(f"Using GeneID {new_gene_id} for stable_id instead of {gene.id}")
-                        return new_gene_id
-
-            # Make a new stable_id
-            if self.make_missing_stable_ids:
-                new_id = self.generate_stable_id()
-                logging.debug(f"New id: {new_gene_id} -> {new_id}")
-                return new_id
-            raise GFFParserError(f"Can't use invalid gene id for {gene}")
-
-        return new_gene_id
-
-    def generate_stable_id(self) -> str:
-        """Returns a new unique gene stable_id with a prefix.
-
-        The id is made up of a prefix and a number, which is auto incremented.
-        Define the prefix with the param "stable_id_prefix",
-        or use the genome organism_abbrev and prepend "TMP_" to it.
-
-        """
-        if self.stable_id_prefix:
-            prefix = self.stable_id_prefix
-        else:
-            if self.genome:
-                org = self.genome.get("BRC4", {}).get("organism_abbrev")
-            if org is None:
-                prefix = "TMP_PREFIX_"
-            else:
-                prefix = "TMP_" + org + "_"
-            self.stable_id_prefix = prefix
-
-        number = self.current_stable_id_number + 1
-        new_id = f"{prefix}{number}"
-        self.current_stable_id_number = number
-
-        return new_id
-
-    def valid_id(self, name: str) -> bool:
-        """Check that the format of a stable id is valid."""
-
-        if not self.validate_gene_id:
-            return True
-
-        min_length = self.min_id_length
-
-        # Trna (from tRNAscan)
-        if re.search(r"^Trna", name):
-            logging.debug(f"Stable ID is a tRNA from tRNA-scan: {name}")
-            return False
-
-        # Coordinates
-        if re.search(r"^.+:\d+..\d+", name):
-            logging.debug(f"Stable id is a coordinate: {name}")
-            return False
-
-        # Special characters
-        if re.search(r"[ |]", name):
-            logging.debug(f"Stable id contains special characters: {name}")
-            return False
-
-        # Min length
-        if len(name) < min_length:
-            logging.debug(f"Stable id is too short (<{min_length}) {name}")
-            return False
-
-        return True
-
-    def normalize_transcript_id(self, gene_id: str, number: int) -> str:
-        """Use a gene ID and a number to make a formatted transcript ID."""
-
-        transcript_id = f"{gene_id}_t{number}"
-        return transcript_id
-
-    def normalize_cds_id(self, cds_id: str) -> str:
-        """Returns a normalised version of the provided CDS ID.
-
-        The normalisation implies to remove any unnecessary prefixes around the CDS ID. However, if
-        the CDS ID is still not proper, an empty string will be returned.
-
-        """
-
-        prefixes = ["cds-", "cds:"]
-        cds_id = self.remove_prefixes(cds_id, prefixes)
-
-        # Special case: if the ID doesn't look like one, remove it
-        # It needs to be regenerated
-        if not self.valid_id(cds_id):
-            cds_id = ""
-
-        return cds_id
-
-    def normalize_pseudogene_cds_id(self, gene: SeqFeature) -> None:
-        """Normalises the CDS ID of the provided pseudogene.
-
-        Ensure CDS from a pseudogene have a proper ID:
-        - Different from the gene
-        - Derived from the gene if it is not proper
-
-        """
-
-        for transcript in gene.sub_features:
-            for feat in transcript.sub_features:
-                if feat.type == "CDS":
-                    feat.id = self.normalize_cds_id(feat.id)
-                    if feat.id in ("", gene.id):
-                        feat.id = f"{transcript.id}_cds"
-                        feat.qualifiers["ID"] = feat.id
-
-    def remove_prefixes(self, identifier: str, prefixes: List[str]) -> str:
-        """Returns the identifier after removing all the prefixes found in it (if any)."""
-        for prefix in prefixes:
-            if identifier.startswith(prefix):
-                identifier = identifier[len(prefix) :]
-        return identifier
